Latest commit (all unsaved work)

+++ Proton/src/Proton/Events/KeyEvent.h
@@ -7,7 +7,7 @@ namespace Proton
 	class KeyEvent : public Event
 	{
 	public:
-		inline int GetKeyCode() const { return m_KeyCode; }
+		inline uint16_t GetKeyCode() const { return m_KeyCode; }
 
 		EVENT_CLASS_CATEGORY(EventCategoryKeyboard | EventCategoryInput)
